requirements.txt
requests
boto3
pygame-ce==2.5.1
pyscroll==2.31
PyTMX==3.32
PyJWT
